drivers/watchdog: wwdg stm32: fix prescaler setup for newer series.
Series STM32G0, STM32G4, STM32H7, STM32L5, STM32WB, STM32WL have
a newer wwdg ip and store the prescaler value not in bits[8:7],
but bits bits[13:11], use the definitions from ll to account for that.

Remove WWDG_PRESCALER_IS_VALID, this is not required as the prescaler
is not calculated, but known valid prescaler values are tested until a
valid counter is found. Additionally WWDG_PRESCALER_IS_VALID is only
defined in the hal, while the rest of the driver only uses ll.

Signed-off-by: Thomas Stranger <thomas.stranger@outlook.com>