drivers/can: add shared_irq support to can_stm32fd driver
This commit adds shared_irq support to can_stm32fd driver, such that
two can instances can be used on stm32g0.

Additinally to the necessary code changes, bindings are added to
st,stm32-fdacn.yaml and the SHARED_IRQ in automatically enabled in KConfig
if the can devices define a shared-irq phandle-array.

Signed-off-by: Thomas Stranger <thomas.stranger@outlook.com>