dts: st: l5: remap peripheral addresses for secure targets
A proposal to restructure stm32l5 dtsi files in order to allow
build of secure targets.
Common peripherals, flash, and ram are defined with non-secure
addresses and secure targets remap these addresses to regions
for secure access.

This requires a change on the hal side.
Pinctrl.dtsi files should not define the pin-controller node with an
address, but simply overwrite the existing:

instead of:
/ {
	soc {
		pinctrl: pin-controller@42020000 {
			adc1_in5_pa0: adc1_in5_pa0 {
				pinmux = <STM32_PINMUX('A', 0, ANALOG)>;
			};

		};

	        ...
	};
};

It is proposed to define the following:
&pinctrl: {
	adc1_in5_pa0: adc1_in5_pa0 {
		pinmux = <STM32_PINMUX('A', 0, ANALOG)>;
	};

        ...
};

Signed-off-by: Thomas Stranger <thomas.stranger@outlook.com>