boards: stm32l5: enable Trustzone configuration for secure targets
This commit enable ARM_TRUSTZONE_M and TRUSTED_EXECUTION_SECURE for
the secure target of stm32l5 boards.
This configration tells the compiler to target the secure state of CMSE.
Will also emit the necessary __ARM_FEATURE_CMSE symbol value which
is needed for stm32cube hal definitions.
Additionally RUNTIME_NMI is enabled such that an NMI hanlder can
be installed with _NmiHandlerSet().

Signed-off-by: Thomas Stranger <thomas.stranger@outlook.com>