doc: develop: getting_started: fedora dnf5 group installation
The group display names don't work with dnf5.
As displaynames depend on the locale, are not unique and can
depend on the locale dnf only supports the group identifiers.

Distributions with dnf5 like RHEL<=10, Fedora<41, and CentOS stream
shouldn't have any issues with the group id, but have not been tested.

Alternatives considered: use the @sytax for groups, but using
the group command makes it more clear that package groups are installed.

Signed-off-by: Thomas Stranger <thomas.stranger@outlook.com>